back to ints

+++ beacon-chain/forkchoice/protoarray/helpers.go
@@ -15,11 +15,11 @@ func computeDeltas(
 	votes []Vote,
 	oldBalances []uint64,
 	newBalances []uint64,
-) ([]uint64, []Vote, error) {
+) ([]int, []Vote, error) {
 	ctx, span := trace.StartSpan(ctx, "protoArrayForkChoice.computeDeltas")
 	defer span.End()
 
-	deltas := make([]uint64, len(blockIndices))
+	deltas := make([]int, len(blockIndices))
 
 	for validatorIndex, vote := range votes {
 		oldBalance := uint64(0)
@@ -47,19 +47,19 @@ func computeDeltas(
 			if ok {
 				// Protection against out of bound, the `nextDeltaIndex` which defines
 				// the block location in the dag can not exceed the total `delta` length.
-				if nextDeltaIndex >= uint64(len(deltas)) {
+				if int(nextDeltaIndex) >= len(deltas) {
 					return nil, nil, errInvalidNodeDelta
 				}
-				deltas[nextDeltaIndex] += newBalance
+				deltas[nextDeltaIndex] += int(newBalance)
 			}
 
 			currentDeltaIndex, ok := blockIndices[vote.currentRoot]
 			if ok {
 				// Protection against out of bound (same as above)
-				if currentDeltaIndex >= uint64(len(deltas)) {
+				if int(currentDeltaIndex) >= len(deltas) {
 					return nil, nil, errInvalidNodeDelta
 				}
-				deltas[currentDeltaIndex] -= oldBalance
+				deltas[currentDeltaIndex] -= int(oldBalance)
 			}
 		}
